# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gigco REST API",
        "contact": {
            "name": "Support Team",
            "email": "support@gigqo.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "api": {},
        "/api/access-tokens": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Получение bearer токена",
                "description": "Получение bearer токена авторизации по email, паролю",
                "operationId": "CreateAuthToken",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|wN9OyREdhoYsZZ94Zyi6V******fePRo4raS96rMc0e51c61"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Удаление текущего bearer токена",
                "description": "Удаление текущего bearer токена текущей авторизации",
                "operationId": "DeleteAuthToken",
                "responses": {
                    "401": {
                        "description": "Response Message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Получение списка книг",
                "description": "Получение списка книг, с учетом фильтров: price={asc/desc}, quantity={asc/desc}, title={keyword}, description={keyword}. Можно применять фильтры все сразу. Фильтр цены важнее чем фильтр количества, при активации обоих фильтров цены и количества, сортировка количества вторична и работает при одинаковой цене",
                "operationId": "GetBooks",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Ключевое слово из названия"
                    },
                    {
                        "name": "desctiption",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Ключевое слово из описания"
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "asc"
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "desc"
                    },
                    {
                        "name": "avatars",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "true"
                    },
                    {
                        "name": "orders",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "asc"
                    },
                    {
                        "name": "rankOrSalesAbove",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "cover_url": {
                                                        "type": "string"
                                                    },
                                                    "price": {
                                                        "type": "string"
                                                    },
                                                    "quantity": {
                                                        "type": "integer"
                                                    },
                                                    "authors": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer"
                                                                },
                                                                "first_name": {
                                                                    "type": "string"
                                                                },
                                                                "last_name": {
                                                                    "type": "string"
                                                                },
                                                                "rank": {
                                                                    "type": "string"
                                                                },
                                                                "avatar_url": {
                                                                    "type": "string"
                                                                },
                                                                "created_at": {
                                                                    "type": "string",
                                                                    "format": "date-time"
                                                                },
                                                                "updated_at": {
                                                                    "type": "string",
                                                                    "format": "date-time"
                                                                },
                                                                "deleted_at": {
                                                                    "type": "string",
                                                                    "nullable": true
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "sales": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string"
                                                },
                                                "last": {
                                                    "type": "string"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "from": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "url": {
                                                                "type": "string",
                                                                "nullable": true
                                                            },
                                                            "label": {
                                                                "type": "string"
                                                            },
                                                            "active": {
                                                                "type": "boolean"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sale": {
            "post": {
                "tags": [
                    "Sale"
                ],
                "summary": "Создание продажи книги",
                "description": "Endpoint для прожажи книги. Создание записи в БД о продаже книги.",
                "operationId": "CreateSale",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "buyer": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "13"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "book_id": {
                                                "type": "integer",
                                                "example": [
                                                    13,
                                                    14,
                                                    15
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Response Message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "buyers_id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "books": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": "19"
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Title"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "Same description"
                                                            },
                                                            "cover_url": {
                                                                "type": "string",
                                                                "example": "/storage/uploads/vf9A30PQr1t4HSpQ7RswYqevs0Aa0hnwHyUBCoaV.png"
                                                            },
                                                            "price": {
                                                                "type": "string",
                                                                "example": "1000"
                                                            },
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": "10"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": "22374"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}